requirements.txt
Pillow
telethon
vkbottle
lottie
requests
